refactored more of the pipeline into the lib crate
