+++ src/cwe_checker_lib/src/checkers/cwe_789.rs
@@ -28,11 +28,11 @@ use crate::abstract_domain::TryToInterval;
 use crate::analysis::pointer_inference::PointerInference;
 use crate::analysis::vsa_results::*;
 use crate::intermediate_representation::*;
+use crate::pipeline::AnalysisResults;
 use crate::utils::log::CweWarning;
 use crate::utils::log::LogMessage;
 use crate::utils::symbol_utils::get_callsites;
 use crate::utils::symbol_utils::get_symbol_map;
-use crate::AnalysisResults;
 use crate::CweModule;
 use serde::Deserialize;
 use serde::Serialize;
