+++ src/cwe_checker_lib/src/utils/mod.rs
@@ -10,14 +10,14 @@ pub mod symbol_utils;
 use crate::prelude::*;
 
 /// Get the contents of a configuration file.
-pub fn read_config_file(filename: &str) -> serde_json::Value {
+pub fn read_config_file(filename: &str) -> Result<serde_json::Value, Error> {
     let project_dirs = directories::ProjectDirs::from("", "", "cwe_checker")
-        .expect("Could not discern location of configuration files.");
+        .context("Could not discern location of configuration files.")?;
     let config_dir = project_dirs.config_dir();
     let config_path = config_dir.join(filename);
     let config_file =
-        std::fs::read_to_string(config_path).expect("Could not read configuration file");
-    serde_json::from_str(&config_file).unwrap()
+        std::fs::read_to_string(config_path).context("Could not read configuration file")?;
+    Ok(serde_json::from_str(&config_file)?)
 }
 
 /// Get the folder path to a Ghidra plugin bundled with the cwe_checker.
